medresearch-ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
langchain==0.2.6
langchain-community==0.2.6
chromadb==0.5.0
sentence-transformers==3.0.0
pypdf==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
ollama==0.2.1
langchain-text-splitters>=0.2.0,<0.3.0
